+++ src/Symfony/Component/Form/Extension/Core/Type/DateType.php
@@ -260,11 +260,10 @@ class DateType extends AbstractType
             'years' => range((int) date('Y') - 5, (int) date('Y') + 5),
             'months' => range(1, 12),
             'days' => range(1, 31),
-            'widget' => function (Options $options) {
-                trigger_deprecation('symfony/form', '6.3', 'Not configuring the "widget" option explicitly is deprecated, its default value will change to "single_text" in 7.0.');
+            'widget' => static function (Options $options) {
+                trigger_deprecation('symfony/form', '6.3', 'Not configuring the "widget" option of form type "date" is deprecated. It will default to "single_text" in Symfony 7.0.');
 
                 return 'choice';
-                // return 'single_text';
             },
             'input' => 'datetime',
             'format' => $format,